requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.8
python-dotenv==1.0.0
gunicorn>=22.0.0,<23.0
psycopg2-binary==2.9.7